migrate loading to inferno
